server/requirements.txt
flask>=2.2.5,<3.0.0
flask-cors>=4.0.0,<5.0.0
pypdf2>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
google-generativeai>=0.4.0,<0.5.0
sentence-transformers>=2.2.2,<3.0.0
werkzeug>=2.2.3,<3.0.0

# Compatibility fixes
numpy>=2.0.0,<3.0.0
pybind11>=2.12.0
scikit-learn>=1.4.0,<2.0.0
transformers>=4.40.0,<5.0.0
